Move Final dataclass-field subsystem into its own module
The typing.Final[T] validation and cache-planning code (~275 lines) had been added to
_dataclass_util.py, which previously held only create_flat_name. Move it verbatim into a
new _final_dataclass_fields.py so the feature is partitioned from the pre-existing shared
utility (per AGENTS.md "minimize contact area"), restoring _dataclass_util.py to its
original contents. Call sites import final_field_names / is_final_annotation from the new
module; create_flat_name imports are unchanged.
