+++ python/quadrants/lang/_template_mapper_hotpath.py
@@ -38,7 +38,8 @@ from quadrants._tensor import (
 )
 from quadrants._tensor_wrapper import _TENSOR_WRAPPER_TYPES
 from quadrants._tensor_wrapper import Tensor as _TensorClass
-from quadrants.lang._dataclass_util import create_flat_name, final_field_names
+from quadrants.lang._dataclass_util import create_flat_name
+from quadrants.lang._final_dataclass_fields import final_field_names
 from quadrants.lang._ndarray import Ndarray
 from quadrants.lang.any_array import AnyArray
 from quadrants.lang.buffer_view import BufferView as BufferViewInstance
